requirements.txt
numpy
tqdm
matplotlib
gym==0.21.0
pygame
ipython
ipywidgets
opencv-python
pandas
pyyaml
scipy
python-dateutil
networkx